Move XmApiError out of the types.ts file
